storing colors as a unsigned byte types (packed into 32-bit signed integers for vertices) and lots of cleanup and very little refactoring

+++ lua/src/script/noise.c
@@ -24,14 +24,14 @@ static int noise_perlin(lua_State *L) {
 
   float value = 0.0F;
 
-  float x = luaL_checknumber(L, 1);
-  float y = luaL_checknumber(L, 2);
-  float z = luaL_checknumber(L, 3);
+  float x = (float)luaL_checknumber(L, 1);
+  float y = (float)luaL_checknumber(L, 2);
+  float z = (float)luaL_checknumber(L, 3);
 
   if (lua_isnoneornil(L, 4)) {
     lua_pushnumber(L, stb_perlin_noise3(x, y, z, 0, 0, 0));
   } else {
-    int seed = luaL_checkinteger(L, 4);
+    int seed = (int)luaL_checkinteger(L, 4);
     lua_pushnumber(L, stb_perlin_noise3_seed(x, y, z, 0, 0, 0, seed));
   }
 
@@ -41,13 +41,13 @@ static int noise_perlin(lua_State *L) {
 static int noise_ridge(lua_State *L) {
   lua_settop(L, 7);
 
-  float x = luaL_checknumber(L, 1);
-  float y = luaL_checknumber(L, 2);
-  float z = luaL_checknumber(L, 3);
-  float lacunarity = luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
-  float gain = luaL_optnumber(L, 5, DEFAULT_GAIN);
-  float offset = luaL_optnumber(L, 6, DEFAULT_OFFSET);
-  int octaves = luaL_optinteger(L, 7, DEFAULT_OCTAVES);
+  float x = (float)luaL_checknumber(L, 1);
+  float y = (float)luaL_checknumber(L, 2);
+  float z = (float)luaL_checknumber(L, 3);
+  float lacunarity = (float)luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
+  float gain = (float)luaL_optnumber(L, 5, DEFAULT_GAIN);
+  float offset = (float)luaL_optnumber(L, 6, DEFAULT_OFFSET);
+  int octaves = (int)luaL_optinteger(L, 7, DEFAULT_OCTAVES);
 
   lua_pushnumber(
       L, stb_perlin_ridge_noise3(x, y, z, lacunarity, gain, offset, octaves));
@@ -58,12 +58,12 @@ static int noise_ridge(lua_State *L) {
 static int noise_fbm(lua_State *L) {
   lua_settop(L, 6);
 
-  float x = luaL_checknumber(L, 1);
-  float y = luaL_checknumber(L, 2);
-  float z = luaL_checknumber(L, 3);
-  float lacunarity = luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
-  float gain = luaL_optnumber(L, 5, DEFAULT_GAIN);
-  int octaves = luaL_optinteger(L, 6, DEFAULT_OCTAVES);
+  float x = (float)luaL_checknumber(L, 1);
+  float y = (float)luaL_checknumber(L, 2);
+  float z = (float)luaL_checknumber(L, 3);
+  float lacunarity = (float)luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
+  float gain = (float)luaL_optnumber(L, 5, DEFAULT_GAIN);
+  int octaves = (int)luaL_optinteger(L, 6, DEFAULT_OCTAVES);
 
   lua_pushnumber(L, stb_perlin_fbm_noise3(x, y, z, lacunarity, gain, octaves));
 
@@ -73,12 +73,12 @@ static int noise_fbm(lua_State *L) {
 static int noise_turbulence(lua_State *L) {
   lua_settop(L, 6);
 
-  float x = luaL_checknumber(L, 1);
-  float y = luaL_checknumber(L, 2);
-  float z = luaL_checknumber(L, 3);
-  float lacunarity = luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
-  float gain = luaL_optnumber(L, 5, DEFAULT_GAIN);
-  int octaves = luaL_optinteger(L, 6, DEFAULT_OCTAVES);
+  float x = (float)luaL_checknumber(L, 1);
+  float y = (float)luaL_checknumber(L, 2);
+  float z = (float)luaL_checknumber(L, 3);
+  float lacunarity = (float)luaL_optnumber(L, 4, DEFAULT_LACUNARITY);
+  float gain = (float)luaL_optnumber(L, 5, DEFAULT_GAIN);
+  int octaves = (int)luaL_optinteger(L, 6, DEFAULT_OCTAVES);
 
   lua_pushnumber(
       L, stb_perlin_turbulence_noise3(x, y, z, lacunarity, gain, octaves));
